--- routers/todos.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const controller = require("../controllers/todos");
-
-router.route("/").post(controller.postTodo);
-
-module.exports = router;
